Multiple bugfixes in AcceleratedReactor, added context for better access to kernel stash

+++ src/lib/XyrusWorx.Runtime.Graphics/IRenderLoop.cs
@@ -12,8 +12,8 @@ namespace XyrusWorx.Runtime.Graphics
 		[CanBeNull]
 		IReactor CurrentReactor { get; }
 		
-		double Clock { get; }
-		double FramesPerSecond { get; }
-		double MaximumFramesPerSecond { get; set; }
+		float Clock { get; }
+		float FramesPerSecond { get; }
+		float MaximumFramesPerSecond { get; set; }
 	}
 }
